+++ src/backend/app/modules/Transaction/Provider/AgentTransactionProvider.php
@@ -39,7 +39,7 @@ class AgentTransactionProvider implements ITransactionProvider {
     private function assignData(array $data): void {
         // provider specific data
         $this->agentTransaction->agent_id = (int) $data['agent_id'];
-        $this->agentTransaction->mobile_device_id = (int) $data['device_id'];
+        $this->agentTransaction->mobile_device_id = $data['device_id'];
 
         // common transaction data
         $this->transaction->amount = (int) $data['amount'];
@@ -169,7 +169,7 @@ class AgentTransactionProvider implements ITransactionProvider {
         return $this->transaction->message;
     }
 
-    public function getAmount(): int {
+    public function getAmount(): float {
         return $this->transaction->amount;
     }
